feat: macro to implement default handle fns on plugins
