+++ rslib/src/handle.rs
@@ -1,3 +1,4 @@
+use crate::base_handle;
 use crate::error::JanusGatewayCommunicationError;
 use crate::protocol::Candidate;
 use crate::protocol::GenericEvent;
@@ -32,89 +33,6 @@ impl Handle {
 
 #[uniffi::export(async_runtime = "tokio")]
 impl Handle {
-    pub async fn fire_and_forget(
-        &self,
-        data: Vec<u8>,
-    ) -> Result<(), JanusGatewayCommunicationError> {
-        let Ok(body) = serde_json::from_slice(&data) else {
-            return Err(JanusGatewayCommunicationError::Serialize {
-                body: String::from_utf8_lossy(&data).to_string(),
-            });
-        };
-        if let Err(why) = self.inner.fire_and_forget(body).await {
-            return Err(JanusGatewayCommunicationError::SendFailure {
-                reason: why.to_string(),
-            });
-        };
-        Ok(())
-    }
-
-    pub async fn fire_and_forget_with_jsep(
-        &self,
-        data: Vec<u8>,
-        jsep: Jsep,
-    ) -> Result<(), JanusGatewayCommunicationError> {
-        let Ok(body) = serde_json::from_slice(&data) else {
-            return Err(JanusGatewayCommunicationError::Serialize {
-                body: String::from_utf8_lossy(&data).to_string(),
-            });
-        };
-        if let Err(why) = self
-            .inner
-            .fire_and_forget_with_jsep(body, jsep.into())
-            .await
-        {
-            return Err(JanusGatewayCommunicationError::SendFailure {
-                reason: why.to_string(),
-            });
-        };
-        Ok(())
-    }
-
-    pub async fn send_waiton_ack(
-        &self,
-        data: Vec<u8>,
-        timeout: Duration,
-    ) -> Result<(), JanusGatewayCommunicationError> {
-        let Ok(body) = serde_json::from_slice(&data) else {
-            return Err(JanusGatewayCommunicationError::Serialize {
-                body: String::from_utf8_lossy(&data).to_string(),
-            });
-        };
-        if let Err(why) = self.inner.send_waiton_ack(body, timeout).await {
-            return Err(JanusGatewayCommunicationError::SendFailure {
-                reason: why.to_string(),
-            });
-        };
-        Ok(())
-    }
-
-    pub async fn send_waiton_result(
-        &self,
-        data: Vec<u8>,
-        timeout: Duration,
-    ) -> Result<Vec<u8>, JanusGatewayCommunicationError> {
-        let Ok(body) = serde_json::from_slice(&data) else {
-            return Err(JanusGatewayCommunicationError::Serialize {
-                body: String::from_utf8_lossy(&data).to_string(),
-            });
-        };
-        let result = match self.inner.send_waiton_rsp::<Value>(body, timeout).await {
-            Ok(result) => result,
-            Err(why) => {
-                return Err(JanusGatewayCommunicationError::SendFailure {
-                    reason: why.to_string(),
-                })
-            }
-        };
-        let Ok(result) = serde_json::from_value(result) else {
-            return Err(JanusGatewayCommunicationError::Serialize {
-                body: String::from_utf8_lossy(&data).to_string(),
-            });
-        };
-        Ok(result)
-    }
-
     pub async fn start_event_loop(&self, cb: Box<dyn HandleCallback>) {
         let Ok(Some(mut receiver)) = self.receiver.lock().map(|mut x| x.take()) else {
             return;
@@ -141,68 +59,9 @@ impl Handle {
             *abort_handle = Some(join_handle.abort_handle());
         }
     }
-
-    pub async fn hangup(&self, timeout: Duration) -> Result<(), JanusGatewayCommunicationError> {
-        self.inner.hangup(timeout).await.map_err(|err| {
-            JanusGatewayCommunicationError::SendFailure {
-                reason: err.to_string(),
-            }
-        })
-    }
-
-    pub async fn detach(&self, timeout: Duration) -> Result<(), JanusGatewayCommunicationError> {
-        self.inner.detach(timeout).await.map_err(|err| {
-            JanusGatewayCommunicationError::SendFailure {
-                reason: err.to_string(),
-            }
-        })
-    }
-
-    pub async fn trickle_single_candidate(
-        &self,
-        candidate: Candidate,
-        timeout: Duration,
-    ) -> Result<(), JanusGatewayCommunicationError> {
-        self.inner
-            .trickle_single_candidate(candidate.into(), timeout)
-            .await
-            .map_err(|err| JanusGatewayCommunicationError::SendFailure {
-                reason: err.to_string(),
-            })
-    }
-
-    pub async fn trickle_candidates(
-        &self,
-        candidates: Vec<Candidate>,
-        timeout: Duration,
-    ) -> Result<(), JanusGatewayCommunicationError> {
-        self.inner
-            .trickle_candidates(candidates.into_iter().map(Into::into).collect(), timeout)
-            .await
-            .map_err(|err| JanusGatewayCommunicationError::SendFailure {
-                reason: err.to_string(),
-            })
-    }
-
-    pub async fn complete_trickle(
-        &self,
-        timeout: Duration,
-    ) -> Result<(), JanusGatewayCommunicationError> {
-        self.inner.complete_trickle(timeout).await.map_err(|err| {
-            JanusGatewayCommunicationError::SendFailure {
-                reason: err.to_string(),
-            }
-        })
-    }
 }
 
-impl Drop for Handle {
-    fn drop(&mut self) {
-        if let Ok(Some(abort_handle)) = self.abort_handle.lock().map(|mut x| x.take()) {
-            abort_handle.abort();
-        }
-    }
-}
+base_handle!(Handle);
 
 #[uniffi::export(callback_interface)]
 pub trait HandleCallback: Send + Sync + Debug {
